% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerTestContrast.R
\name{lmerTestContrast}
\alias{lmerTestContrast}
\title{lmerTestContrast}
\usage{
lmerTestContrast(fm, contrast, df_prior = 0, s2_prior = 0)
}
\arguments{
\item{fm}{\itemize{
\item linear mixed model fit by \code{lmerTest::lmer}.
}}

\item{contrast}{\itemize{
\item a vector indicating a contrast between model coefficients.
See the \code{getContrast} function to create a contrast.
}}

\item{df_prior}{\itemize{
\item prior degrees of freedom for moderated comparison; see also
\code{limma::squeezeVar} to compute prior degrees of freedom.
}}

\item{s2_prior}{\itemize{
\item prior sigma squared for moderated comparison; see also
\code{limma::squeezeVar} to compute prior sigma squared.
}}
}
\description{
evaluate model-based comparison given a linear mixed-model as fit by lmerTest
}
