% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmTestContrast.R
\name{lmTestContrast}
\alias{lmTestContrast}
\title{lmTestContrast}
\usage{
lmTestContrast(fm, LT, s2_prior = 0, df_prior = 0)
}
\arguments{
\item{fm}{\itemize{
\item fit linear model
}}

\item{s2_prior}{\itemize{
\item prior sigma squared for moderated comparison; see also
}}

\item{df_prior}{\itemize{
\item prior degrees of freedom for moderated comparison; see also
\code{limma::squeezeVar} to compute prior degrees of freedom.
}}

\item{contrast}{\itemize{
\item a vector indicating a contrast between model coefficients.
See the \code{getContrast} function to create a contrast.
}}
}
\value{
a data.table giving the result of the model-based comparison,
contains the following columns:
\itemize{
\item{Contrast - }{A vector defining a comparison between positive
and negative coefficients in a model. See also, \code{getContrast}.}
\item{log2FC - }{The fold change estimated from the fit model.}
\item{percentControl}{The fold-change converted to a percentage of the
control.}
\item{SE - }{The standard error of the comparison computed as the square
root of the variance.}
\item{Tstatistic - }{The t-statistic for the comparison, calculated using
equation 11 from Kuznetsova et al., 2017.}
\item{Pvalue - }{The p-value for the comparison, calculated from the t-value
and degrees of freedom using the student's t-distribution, \code{pt}.}
\item{DF - }{The degrees of freedom for the comparison.}
\item{S2 - }{sigma squared -- the estimated standard deviation of the errors
or residual standard deviation (sigma) squared -- see also \code{?sigma}.}
}
\description{
assess a contrast for a given linear model as fit by lm
}
